+++ packages/utils/src/index.ts
@@ -1,6 +1,6 @@
-import memorize from './memorize'
 import * as random from './random'
-import * as math from './math'
+import memorize from './memorize'
+import math from './math'
 
 export {
   random,