+++ packages/web-stg/src/library.js
@@ -1,4 +1,3 @@
-import '@stg/core/dist/index.min.css'
 import { Bullet } from '@stg/core'
 import * as builtin from '@stg/bullets'
 
@@ -6,4 +5,4 @@ Bullet.install(builtin)
 
 export * from '@stg/core'
 export * from '@stg/utils'
-export const version = '0.2.1'
+export { version } from '../package.json'