+++ packages/web-stg/src/runtime.js
@@ -1,0 +1,2 @@
+export * from './library'
+import '../styles/index.scss'