+++ packages/docs/guide/CHANGELOG.md
@@ -1,21 +1,33 @@
 # 更新日志
 
+## v0.3.1
+
+### 新特性
+
+- **core:** 新增 `loadImages` 和 `loadAssets` 方法, 用于加载全局资源.
+
+### 修复与其他
+
+打包时将生成多个入口文件:
+
+- stg.min.css: 所需的样式文件.
+- stg.common.js: 可直接在 NodeJS 环境下加载的模块.
+- stg.runtime.js: 应当在浏览器中加载使用的模块, 内置了所需的样式.
+
 ## v0.3
 
 ### 重大变化
 
-本版本做出了一些变化, 目的是让核心模块与浏览器彻底解耦. 依赖浏览器环境的部分将放入 web-stg 包中.
-
 - **core:** Looping 不再是 EventTarget 的子类, 对应的事件全部改用 `onPause` 等属性来实现.
 - **core:** Looping 类的 onUpdate 事件更改为 onStats, 同时增加 `statsInterval` 属性用于控制其触发频率.
-- **core:** Looping 构造函数新增 `skipFrame` 参数, 可用于调整描画间隔.
-- **core:** 新增 `looping.initialize()` 方法, 可以用于重新初始化实例.
 - **core:** 移除 `looping.resume()` 等方法中的 `slient` 参数.
-- **core:** `looping.getStatus()` 返回对象新增 `frameRate` 属性, 表示当前物理帧数.
 
 ### 新特性
 
+- **core:** Looping 构造函数新增 `skipFrame` 参数, 可用于调整描画间隔.
+- **core:** 新增 `looping.initialize()` 方法, 可以用于重新初始化实例.
 - **core:** 新增 `core.config.angleUnit`, 用于配置采用的角度单位, 默认为 `Math.PI`.
+- **core:** `looping.getStatus()` 返回对象新增 `frameRate` 属性, 表示当前物理帧数.
 
 ## v0.2.1